Map all virtual memory to the guest on start
Right now, we map virtual memory regions on demand by handling KVM_EXIT_MMIO
events. The reason is that we have a limited number of memory slots that can't
cover the whole guest physical address space.

But I think it is a time bomb. If we don't limit the guest address space,
we can end up in a situation when we don't have memory slots to map a new
region and we will panic. This can happen if we create many small mappings.

This change limits the guest physical address space to the size that can be
handled with available memory slots. Here are a few benefits:
* We can map the sentry virtual address space to the guest on start and remove
  the code that handles KVM_EXIT_MMIO.
* We can be sure that the sentry will use the whole available address space.

PiperOrigin-RevId: 391558063