Demonstrate S/R end-case with raw IP sockets
Add a syscall tests that demonstrates that when a packet is received, a
packet buffer is delivered to all interested transport endpoints without
making a deep copy of any views. When a save is triggered with multiple
endpoints holding a reference to the same underlying view, S/R panics.

In the added test, the raw socket holds a reference to the full UDP
packet (which is held in a single view) but the UDP socket only holds
a reference to the UDP payload (which is an alias + offset to the view
that the raw socket holds).

Note, S/R does not support aliasing.

PiperOrigin-RevId: 393177049